requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0